requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0

# LoRA and PEFT
peft>=0.4.0
loralib>=0.1.1

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# Data Science and Utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Configuration and Logging
omegaconf>=2.3.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Web and Streaming
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0
streamlit>=1.25.0
gradio>=3.35.0

# Async and Queue Management
asyncio-mqtt>=0.13.0
redis>=4.6.0
celery>=5.3.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
openvino>=2023.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=6.5.0

# System and Environment
python-dotenv>=1.0.0
psutil>=5.9.0
GPUtil>=1.4.0
